Fixed Style

+++ tasks/rational/rational.cpp
@@ -40,7 +40,7 @@ Rational& operator+=(Rational& lhs, const Rational& rhs) {
     int64_t new_numer = static_cast<int64_t>(lhs.GetDenominator()) * static_cast<int64_t>(rhs.GetNumerator()) +
                         static_cast<int64_t>(rhs.GetDenominator()) * static_cast<int64_t>(lhs.GetNumerator());
     int64_t new_denom = static_cast<int64_t>(lhs.GetDenominator()) * static_cast<int64_t>(rhs.GetDenominator());
-    int gcd = std::gcd(new_denom, new_numer);
+    int64_t gcd = std::gcd(new_denom, new_numer);
     new_numer /= gcd;
     new_denom /= gcd;
     lhs.Set(new_numer, new_denom);
@@ -50,7 +50,7 @@ Rational& operator+=(Rational& lhs, const Rational& rhs) {
 Rational& operator*=(Rational& lhs, const Rational& rhs) {
     int64_t new_numer = static_cast<int64_t>(lhs.GetNumerator()) * static_cast<int64_t>(rhs.GetNumerator());
     int64_t new_denom = static_cast<int64_t>(lhs.GetDenominator()) * static_cast<int64_t>(rhs.GetDenominator());
-    int gcd = std::gcd(new_denom, new_numer);
+    int64_t gcd = std::gcd(new_denom, new_numer);
     new_numer /= gcd;
     new_denom /= gcd;
     lhs.Set(new_numer, new_denom);
